Incorporated comments from the dreaded reviewer #2.
III-A, par. 2: What does it mean for the "representation of the expression 
to be evaluated" to be "unbiased"? 

III-B: This section is very confusing as written. First a random variable is 
defined as a 4-tuple, then an expression is presented which looks nothing 
like a 4-tuple ([X => Normal(mu, sigma^2)]), then there is an assertion that 
random variables are stored "as arithmetic formulas". 

III-B An example of how "bag semantics [are used] to encode disjunctions" 
would be nice. 

Algorithm 3.2: This algorithm is hard to read in the absence of any comments 
explaining what is going on. The notation S[Y,Z,...] is never defined, nor 
is "interval map", nor are the notions of "strong" versus "weak" 
consistency. 

IV: The text below equation (1) looks as if it may contain a redundant 
paragraph. 

IV: The phrase "or about 0.17" is unclear as written. 

IV-A: The authors should give a reference for the method of inverse CDFs as 
used to sample conditional distribution (E.g., Averill Law's simulation 
textbook). 

Algorithm 4.2: This algorithm is extremely hard to read. The inputs and 
outputs are never clearly defined. (E.g., the result is a pair of values, 
but what do they mean? What is the interpretation of epsilon and delta?) 
There are undefined terms such as "erf" and "Natural" and P[K]. There are no 
comments whatsoever---it appears that a central-limit-theorem based 
probabilisitic error bound is being used, but this is never explained 
clearly; please keep the ICDE audience in mind. 

IV-C: Fix the sentence fragment just below Algorithm 4.2. 

IV-C: in the complete paragraph below Algorithm 4.2, it is incorrectly 
stated that the sum of N random variables, each with variance sigma, has a 
variance of sigma/sqrt(N). The AVERAGE of N variables with STANDARD 
DEVIATION sigma has STANDARD DEVIATION sigma/sqrt(N). As a consequence of 
these typos, the subsequent discussion is confusing. 

IV-C: Nonstandard angle-bracket notation is used without definition. 

IV-C: The statement "the aggregate must be designed with expectation 
computation in mind" is unclear. Also, what does this mean with regards to 
the scope of applicability of PIP? The discussion of max is hard to follow 
without an example; the phrase "target expression is a constant" is unclear. 

IV-C: In the sentence under Figure 3, change "table table" to "table". 

V-A: The paragraph starting with "To complete this illusion" (illusion?) is 
hard to read, especially the sentence "if either (A>B) resolves to a 
CTYPE...". What does this mean? There is no variable A or B in the rewritten 
query. In general, a few more examples of rewritten queries would help 
greatly in understanding the "seamlessly" assertion at the end of the 
sentence. 

V-C: The last paragraph of this sentence is unclear. What exactly is a 
"composite aggregate variable"? I sithis a symbolic representation of the 
aggregate? Exacyly why would it be useful to maintain one? 

VI, par. 6: Change "queryies" to "queries" 



git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@358 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 3bce40081d73e9b5762ff7e808022ce09f181b30