Finalized the move to near-versionlessness.  (Well, not quite... versions are still required for snapshotting, massput synchronization, and ensuring that gets can be assigned to their corresponding puts... but there's no complicated out of order execution problems due to the versioning).
Continued debugging.  Starting to rewrite the simple regression test to accomodate the new semantics.

Added a simple schema validator to ensure the Thrift service schema is in sync with the corresponding Ruby handler.
Added a handful of simple core mixins to make hash creation and array looping easier.
Added a handful of simple spread_types mixins
  - Entry.make... why thrift doesn't do this automatically, I have no idea.
  - PutParams, a class designed to simplify encoding/decoding PutFields - UNION structs are a pain to encode.  
  - Moved MapNodeInterface to where it belongs, a mixin.  See MapNode::Client.connect and close.



git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@421 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 2391c7f709ef8b7e98c90ab457f22312f66bf513