re: #10 Ok... this is going to be a little more challenging than expected due to more alpha4 calculus stupidity.  extract_named_aggregates assumes that there are multiple levels of nested aggsum, with the outer aggsum referencing the bigsum maps, and the inner aggsum being the thing that we want to extract.  Ugh.
The code for this post processing step is getting checked in, but is hidden behind '-d factor-postprocess'.   Getting this step working correctly is going to require a rewrite of extract_named_aggregates, so unless we identify this as a serious speed impediment, I'm going to push this issue back to alpha5 (when it should be automatically addressed--since factorize-polynomial will be far simpler).



git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@1357 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 5e89c1bfc7d508ee461028267f17c5e73db0cc64