Bugfix for optimization order. IVC/update migration. Imperative stage support for nullification via valid symbol.
We cannot actually nullify any of our queries due to our current factorization strategy, thus I am not checking in the nullify code for K3. The IVC migration is still generally useful to define an update normal form, and we will need the imperative stage support when we do choose to implement nullification.

git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@1399 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 8f13c584698ffcdb00cc81d37116461372d6860d