Multiple issues fixed:
- Providing the correct scope when descending into AggSum, Lifts, Externals, and Exists. Before, the scope would remain intact when recurring into these values causing problems with variable scoping. For example,

  AggSum([A,B], S(A,B) * (X ^= AggSum([A], R(A,B) * {B > A})) * {X > 0}

The problem was that the lifted subexpression had the scope [A,B], which is wrong -- it should have been only [A], B is discarded. This bug would allow the optimizer, for example, to commute the two terms in the lifted subexpression and possibly do some unification.

- Related to this issues, unification was not respecting the scope of variables.

((B ^= 4) * S(B) * AggSum([],((R(A, B) * B)))) would be translated into ((B ^= 4) * S(B) * AggSum([],((R(A, B) * 4)))).

This problem was affecting TPCH-8, for example, and also new test cases (simple/r_case.sql and simple/r_case2.sql). The issue is fixed by extending Calculus.rewrite with one additional parameter (function) which decides whether to descend into values. For unification, we descend only when the lift variable appears in the input or output variables of the expression we are unifying with. 

- Added new test cases to test for the above problems.

git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@3237 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: d3cd7b3f596978572240942ecc6770d1253e07f2