A problem might appear when we materialize expressions joining three or more relations with N-1 and 1-N relationships, which yields a potentially huge N x N result (e.g., observe the joins around NATION in the M3 programs of TPCH5 and TPCH7).
To prevent that, we provide the HEURISTICS-DECOMPOSE-OVER-TABLES flag that decomposes expressions in which one static table joins with two or more streams. The rationale is that static tables are often small, dimension tables referenced through foreign keys of the stream tables. Joining one such static table with two or more streams often results in a quadratic number of output tuples.



git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@3385 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: dc432d72ea94fe260a9240444bbc4284c973cd22